refactor(actions/auth): split into co-located action + test modules
